Organize feature/confirm per concept v1
Move confirm domain models (ConfirmError, ConfirmState, ConfirmProperty,
AmountUIModel, FeeUIModel, FeeRateUIModel) from feature module to
gemcore/domains/confirm.

Extract business logic from ConfirmViewModel into coordinator interfaces
(ValidateBalance, ConfirmTransaction, BuildConfirmProperties) with
implementations in data/coordinators and Hilt DI wiring.

ConfirmViewModel now injects coordinator interfaces instead of raw
repositories and operators, reducing it from 622 to 338 LOC.

Closes #18
