+++ android/gemcore/src/main/kotlin/com/gemwallet/android/application/confirm/coordinators/BuildConfirmProperties.kt
@@ -1,0 +1,12 @@
+package com.gemwallet.android.application.confirm.coordinators
+
+import com.gemwallet.android.domains.confirm.ConfirmProperty
+import com.gemwallet.android.model.AssetInfo
+import com.gemwallet.android.model.ConfirmParams
+
+interface BuildConfirmProperties {
+    suspend operator fun invoke(
+        request: ConfirmParams,
+        assetsInfo: List<AssetInfo>,
+    ): List<ConfirmProperty>
+}
