+++ android/gemcore/src/main/kotlin/com/gemwallet/android/domains/confirm/ConfirmError.kt
@@ -1,29 +1,18 @@
-package com.gemwallet.android.features.confirm.models
+package com.gemwallet.android.domains.confirm
 
 import com.wallet.core.primitives.Asset
 import com.wallet.core.primitives.Chain
 
 sealed class ConfirmError : Exception() {
-
-    object None : ConfirmError()
-
-    object Init : ConfirmError()
-
-    object PreloadError : ConfirmError()
-
-    object TransactionIncorrect : ConfirmError()
-
-    object RecipientEmpty : ConfirmError()
-
+    data object None : ConfirmError()
+    data object Init : ConfirmError()
+    data object PreloadError : ConfirmError()
+    data object TransactionIncorrect : ConfirmError()
+    data object RecipientEmpty : ConfirmError()
+    data object SignFail : ConfirmError()
     class InsufficientBalance(val chainTitle: String) : ConfirmError()
-
     class InsufficientFee(val chain: Chain) : ConfirmError()
-
     class MinimumAccountBalanceTooLow(val asset: Asset, val required: Long) : ConfirmError()
-
-    object SignFail : ConfirmError()
-
     class BroadcastError(val details: String) : ConfirmError()
-
     class DustThreshold(val chainTitle: String) : ConfirmError()
 }
