+++ android/gemcore/src/main/kotlin/com/gemwallet/android/domains/confirm/FeeUIModel.kt
@@ -1,14 +1,12 @@
-package com.gemwallet.android.features.confirm.models
+package com.gemwallet.android.domains.confirm
 
 import com.wallet.core.primitives.Asset
 import com.wallet.core.primitives.FeePriority
 import java.math.BigInteger
 
 sealed interface FeeUIModel {
-    object Calculating : FeeUIModel
-
-    object Error : FeeUIModel
-
+    data object Calculating : FeeUIModel
+    data object Error : FeeUIModel
     class FeeInfo(
         val amount: BigInteger,
         val cryptoAmount: String,
